# Chapter13/FirstImplementation/contracts/openapi/cooking-assistance.openapi.yaml
openapi: 3.1.0
info:
  title: Cooking Assistance API
  description: |
    Provides help for cooks in planning and preparation of meals and dinners.
    Cooks create help requests and select who can answer them; help providers are the community, the Grandma Avatar (AI), and chefs.
  version: 1.0.0
  contact:
    name: Annegret Junker
    email: annegret.junker@larder.org
  x-api-id: 56ee84d9-f6c0-4ad9-92a4-1a4e0a7ba82e
  x-audience: external-public
servers:
  - url: 'https://larder.org/cooking-assistance'
security:
  - oAuth2:
      - help:read
      - help:write
...
paths:
  /help-requests:
    get:
      operationId: listHelpRequests
      tags:
        - Help Requests
      summary: List help requests
      description: |
        Retrieve help requests, optionally filtered by the requesting cook and by status.
        Help providers use the status filter to find open requests they can answer.
      parameters:
        - $ref: '#/components/parameters/RequesterParameter'
        - $ref: '#/components/parameters/StatusParameter'
        - $ref: '#/components/parameters/VersionParameter'
      responses:
        '200':
          $ref: '#/components/responses/HelpRequestsResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          $ref: '#/components/responses/ForbiddenResponse'
        '500':
          $ref: '#/components/responses/ServiceNotAvailableResponse'
        default:
          $ref: '#/components/responses/UnexpectedErrorResponse'

    post:
      operationId: createHelpRequest
      tags:
        - Help Requests
      summary: Create a new help request
      description: |
        Create a help request. The authenticated cook becomes the requester and the request starts in status Open.
        Invariants enforced (violations are answered with 400):
        - a recipe is mandatory for the types Preparation Step Explanation and Ingredient Substitute;
        - a howToStep is only allowed for the type Preparation Step Explanation;
        - ingredients are only allowed for the type Ingredient Substitute;
        - chef support is only available for the type Menu proposal, and chef help is provided exclusively,
          i.e. Chef cannot be combined with another preferred provider.
      security:
        - oAuth2:
            - help:write
      parameters:
        - $ref: '#/components/parameters/VersionParameter'
      requestBody:
        $ref: '#/components/requestBodies/HelpRequestCreateRequest'
      responses:
        '201':
          $ref: '#/components/responses/HelpRequestLinkResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          $ref: '#/components/responses/ForbiddenResponse'
        '500':
          $ref: '#/components/responses/ServiceNotAvailableResponse'
        default:
          $ref: '#/components/responses/UnexpectedErrorResponse'
  /help-requests/{helpRequestId}:
    get:
      operationId: getHelpRequestById
      tags:
        - Help Requests
      summary: Get help request by ID
      description: Retrieve a single help request by its unique identifier.
      parameters:
        - $ref: '#/components/parameters/HelpRequestIdParameter'
        - $ref: '#/components/parameters/VersionParameter'
      responses:
        '200':
          $ref: '#/components/responses/HelpRequestResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          $ref: '#/components/responses/ForbiddenResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/ServiceNotAvailableResponse'
        default:
          $ref: '#/components/responses/UnexpectedErrorResponse'

    patch:
      operationId: updateHelpRequest
      tags:
        - Help Requests
      summary: Change a help request
      description: |
        Partially update a help request; omitted fields stay unchanged.
        Only the requester is allowed to change a help request (403 otherwise).
        The type-dependent invariants of the help request are re-checked against the resulting state.
        Setting the status to Answered is rejected unless at least one help answers the request.
      security:
        - oAuth2:
            - help:write
      parameters:
        - $ref: '#/components/parameters/HelpRequestIdParameter'
        - $ref: '#/components/parameters/VersionParameter'
      requestBody:
        $ref: '#/components/requestBodies/HelpRequestUpdateRequest'
      responses:
        '200':
          $ref: '#/components/responses/HelpRequestLinkResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          $ref: '#/components/responses/ForbiddenResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/ServiceNotAvailableResponse'
        default:
          $ref: '#/components/responses/UnexpectedErrorResponse'

    delete:
      operationId: deleteHelpRequest
      tags:
        - Help Requests
      summary: Delete a help request
      description: |
        Withdraw a help request. This operation is irreversible.
        Only the requester is allowed to delete a help request (403 otherwise),
        and only while the request is in status Open (400 otherwise) - an answered request cannot be deleted.
      security:
        - oAuth2:
            - help:write
      parameters:
        - $ref: '#/components/parameters/HelpRequestIdParameter'
        - $ref: '#/components/parameters/VersionParameter'
      responses:
        '204':
          description: Help request deleted successfully
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          $ref: '#/components/responses/ForbiddenResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/ServiceNotAvailableResponse'
        default:
          $ref: '#/components/responses/UnexpectedErrorResponse'
  /helps:
    get:
      operationId: listHelps
      tags:
        - Helps
      summary: List helps
      description: |
        Retrieve helps, optionally filtered by the help request they answer, by type and by help provider.
        If the help provider is a Chef, only the original requester can see the help;
        such helps are left out of the list for every other caller.
      parameters:
        - $ref: '#/components/parameters/HelpRequestIdQueryParameter'
        - $ref: '#/components/parameters/TypeParameter'
        - $ref: '#/components/parameters/HelpProviderParameter'
        - $ref: '#/components/parameters/VersionParameter'
      responses:
        '200':
          $ref: '#/components/responses/HelpsResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          $ref: '#/components/responses/ForbiddenResponse'
        '500':
          $ref: '#/components/responses/ServiceNotAvailableResponse'
        default:
          $ref: '#/components/responses/UnexpectedErrorResponse'

    post:
      operationId: createHelp
      tags:
        - Helps
      summary: Create a new help
      description: |
        Answer a help request. The help provider is taken from the authenticated caller;
        it is absent when the Grandma Avatar answers.
        Invariants enforced (violations are answered with 400):
        - the answerType must be the same as the type of the help request;
        - recipe, howToStep and ingredients in the answer must be the same as in the help request;
        - a cook can answer their own request - except for Chef requests.
      security:
        - oAuth2:
            - help:write
      parameters:
        - $ref: '#/components/parameters/VersionParameter'
      requestBody:
        $ref: '#/components/requestBodies/HelpCreateRequest'
      responses:
        '201':
          $ref: '#/components/responses/HelpLinkResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          $ref: '#/components/responses/ForbiddenResponse'
        '500':
          $ref: '#/components/responses/ServiceNotAvailableResponse'
        default:
          $ref: '#/components/responses/UnexpectedErrorResponse'
  /helps/{helpId}:
    get:
      operationId: getHelpById
      tags:
        - Helps
      summary: Get help by ID
      description: |
        Retrieve a single help by its unique identifier.
        If the help provider is a Chef, only the original requester can see the help (403 otherwise).
      parameters:
        - $ref: '#/components/parameters/HelpIdParameter'
        - $ref: '#/components/parameters/VersionParameter'
      responses:
        '200':
          $ref: '#/components/responses/HelpResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          $ref: '#/components/responses/ForbiddenResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/ServiceNotAvailableResponse'
        default:
          $ref: '#/components/responses/UnexpectedErrorResponse'

    patch:
      operationId: updateHelp
      tags:
        - Helps
      summary: Change a help
      description: |
        Partially update a help; omitted fields stay unchanged.
        Only the provider of the help is allowed to change it (403 otherwise).
        A replaced answer must keep the answerType of the help request.
      security:
        - oAuth2:
            - help:write
      parameters:
        - $ref: '#/components/parameters/HelpIdParameter'
        - $ref: '#/components/parameters/VersionParameter'
      requestBody:
        $ref: '#/components/requestBodies/HelpUpdateRequest'
      responses:
        '200':
          $ref: '#/components/responses/HelpLinkResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          $ref: '#/components/responses/ForbiddenResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/ServiceNotAvailableResponse'
        default:
          $ref: '#/components/responses/UnexpectedErrorResponse'

    delete:
      operationId: deleteHelp
      tags:
        - Helps
      summary: Delete a help
      description: |
        Remove a help. This operation is irreversible.
        Only the provider of the help is allowed to delete it (403 otherwise).
      security:
        - oAuth2:
            - help:write
      parameters:
        - $ref: '#/components/parameters/HelpIdParameter'
        - $ref: '#/components/parameters/VersionParameter'
      responses:
        '204':
          description: Help deleted successfully
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          $ref: '#/components/responses/ForbiddenResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/ServiceNotAvailableResponse'
        default:
          $ref: '#/components/responses/UnexpectedErrorResponse'
components:
  parameters:
    HelpRequestIdParameter:
      name: helpRequestId
      in: path
      required: true
      description: Unique identifier of the help request.
      schema:
        type: string
        format: uuid
    HelpIdParameter:
      name: helpId
      in: path
      required: true
      description: Unique identifier of the help.
      schema:
        type: string
        format: uuid
    RequesterParameter:
      name: requester
      in: query
      required: false
      description: Return only help requests raised by this cook.
      schema:
        type: string
        format: uuid
    StatusParameter:
      name: status
      in: query
      required: false
      description: Return only help requests in this status.
      schema:
        $ref: '#/components/schemas/HelpRequestStatus'
    HelpRequestIdQueryParameter:
      name: helpRequestId
      in: query
      required: false
      description: Return only helps that answer this help request.
      schema:
        type: string
        format: uuid
    TypeParameter:
      name: type
      in: query
      required: false
      description: Return only helps whose answer is of this type.
      schema:
        $ref: '#/components/schemas/HelpType'
    HelpProviderParameter:
      name: helpProvider
      in: query
      required: false
      description: Return only helps given by this help provider.
      schema:
        type: string
        format: uuid
    VersionParameter:
      name: version
      in: header
      required: true
      description: Version of the API the client was built against.
      schema:
        type: string
        default: "1.0.0"
        enum: ["1.0.0"]
  schemas:
    HelpType:
      description: |
        Kind of help a cook asks for; a help answers with the same kind (answerType).
      type: string
      enum:
        - INGREDIENT_SUBSTITUTE
        - PREPARATION_STEP_EXPLANATION
        - STEPS_TO_MITIGATE_CATASTROPHE
        - MENU_PROPOSAL
...
    HelpRequestStatus:
      description: |
        Status of a help request. A request starts as Open; it can only be Answered when at least one help answers it.
      type: string
      enum:
        - OPEN
        - ANSWERED
...
    HelpRequest:
      type: object
      required:
        - helpRequestId
        - requester
        - title
        - type
        - description
        - preferredProvider
        - status
      allOf:
        - $ref: '#/components/schemas/HelpRequestInvariants'
      properties:
        helpRequestId:
          description: Unique identifier of the help request, generated by the server.
          type: string
          format: uuid
          examples:
            - 23a8eeed-35f6-460b-892e-7bb458a8fded
        requester:
          description: The cook who raised the help request, taken from the authenticated caller.
          type: string
          format: uuid
          examples:
            - f64e07f0-f9e7-4b7a-8695-e7a8b3ef5074
        title:
          description: Short headline telling help providers what the request is about.
          type: string
          minLength: 1
          maxLength: 200
          examples:
            - Burning Catastrophe
        type:
          description: Kind of help the cook asks for; decides which further fields are allowed and which answer is expected.
          $ref: '#/components/schemas/HelpType'
        description:
          description: The cook's own account of the situation and what went wrong.
          type: string
          minLength: 1
          maxLength: 5000
          examples:
            - Scones are burned and mother in law is coming in 30 minutes
        recipe:
          description: |
            Reference to the recipe the request is about. Mandatory for the types
            Preparation Step Explanation and Ingredient Substitute.
          type: string
          format: uuid
          examples:
            - 7cf09822-77a1-46bb-812f-b7852bca0913
        howToStep:
          description: Reference to the unclear step of the recipe. Only for the type Preparation Step Explanation.
          type: string
          format: uuid
          examples:
            - 65610dee-fb83-4341-a730-26a4a99a621a
        ingredients:
          description: References to the ingredients to be substituted. Only for the type Ingredient Substitute.
          $ref: '#/components/schemas/IngredientIds'
        preferredProvider:
          description: Who the requester selects to answer the request.
          $ref: '#/components/schemas/PreferredProviders'
        status:
          description: Whether the request is still open or has been answered by at least one help.
          $ref: '#/components/schemas/HelpRequestStatus'
        createdAt:
          description: Timestamp of creation, set by the server.
          type: string
          format: date-time
        updatedAt:
          description: Timestamp of the last change, set by the server.
          type: string
          format: date-time
    HelpRequests:
      description: List of help requests.
      type: array
      items:
        $ref: '#/components/schemas/HelpRequest'
    HelpRequestCreate:
      type: object
      required:
        - title
        - type
        - description
        - preferredProvider
      allOf:
        - $ref: '#/components/schemas/HelpRequestInvariants'
      properties:
        title:
          description: Short headline telling help providers what the request is about.
          type: string
          minLength: 1
          maxLength: 200
          examples:
            - Burning Catastrophe
        type:
          description: Kind of help the cook asks for; decides which further fields are allowed and which answer is expected.
          $ref: '#/components/schemas/HelpType'
        description:
          description: The cook's own account of the situation and what went wrong.
          type: string
          minLength: 1
          maxLength: 5000
        recipe:
          description: |
            Reference to the recipe the request is about. Mandatory for the types
            Preparation Step Explanation and Ingredient Substitute.
          type: string
          format: uuid
        howToStep:
          description: Reference to the unclear step of the recipe. Only for the type Preparation Step Explanation.
          type: string
          format: uuid
        ingredients:
          description: References to the ingredients to be substituted. Only for the type Ingredient Substitute.
          $ref: '#/components/schemas/IngredientIds'
        preferredProvider:
          description: Who the requester selects to answer the request.
          $ref: '#/components/schemas/PreferredProviders'
    HelpRequestUpdate:
      type: object
      properties:
        title:
          description: Short headline telling help providers what the request is about.
          type: string
          minLength: 1
          maxLength: 200
        description:
          description: The cook's own account of the situation and what went wrong.
          type: string
          minLength: 1
          maxLength: 5000
        recipe:
          description: Reference to the recipe the request is about; null removes it where the type allows.
          type: [string, "null"]
          format: uuid
        howToStep:
          description: Reference to the unclear step of the recipe; null removes it. Only for the type Preparation Step Explanation.
          type: [string, "null"]
          format: uuid
        ingredients:
          description: References to the ingredients to be substituted. Only for the type Ingredient Substitute.
          $ref: '#/components/schemas/IngredientIds'
        preferredProvider:
          description: Who the requester selects to answer the request.
          $ref: '#/components/schemas/PreferredProviders'
        status:
          description: New status of the request; Answered is only accepted when at least one help answers it.
          $ref: '#/components/schemas/HelpRequestStatus'
...
    Helps:
      description: List of helps.
      type: array
      items:
        $ref: '#/components/schemas/Help'
    HelpCreate:
      type: object
      required:
        - helpRequest
        - answerTitle
        - helpProviderType
        - answer
      properties:
        helpRequest:
          description: Reference to the help request this help answers.
          type: string
          format: uuid
        answerTitle:
          description: Short headline of the answer.
          type: string
          minLength: 1
          maxLength: 200
        helpProviderType:
          description: Whether a chef, the Grandma Avatar or the community gives this help.
          $ref: '#/components/schemas/HelpProviderType'
        answer:
          description: The content of the help; its answerType must be the same as the request type.
          $ref: '#/components/schemas/Answer'
    HelpUpdate:
      type: object
      properties:
        answerTitle:
          description: Short headline of the answer.
          type: string
          minLength: 1
          maxLength: 200
        answer:
          description: Replacement content of the help; its answerType must stay the same as the request type.
          $ref: '#/components/schemas/Answer'
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          description: Machine-readable error code.
          type: string
        message:
          description: Human-readable explanation of the error.
          type: string
        details:
          description: Additional, error-specific information, e.g. the violated invariant.
          type: object
  responses:
    HelpRequestResponse:
      description: Single help request response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/HelpRequest'
          examples:
            burningCatastrophe:
              $ref: '#/components/examples/BurningCatastropheHelpRequest'
    HelpRequestsResponse:
      description: List of help requests response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/HelpRequests'
          examples:
            openRequests:
              $ref: '#/components/examples/HelpRequestList'
    HelpRequestLinkResponse:
      description: Link to the created or changed help request
      content:
        application/json:
          schema:
            title: HelpRequestLink
            type: object
            required:
              - helpRequestLink
            properties:
              helpRequestLink:
                description: Link to the created or changed help request.
                type: string
                format: uri
    HelpResponse:
      description: Single help response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Help'
          examples:
            stayCalm:
              $ref: '#/components/examples/StayCalmHelp'
    HelpsResponse:
      description: List of helps response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Helps'
    HelpLinkResponse:
      description: Link to the created or changed help
      content:
        application/json:
          schema:
            title: HelpLink
            type: object
            required:
              - helpLink
            properties:
              helpLink:
                description: Link to the created or changed help.
                type: string
                format: uri
    BadRequestResponse:
      description: Invalid request, including violated invariants
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UnauthorizedResponse:
      description: Authentication missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ForbiddenResponse:
      description: Access forbidden, e.g. the caller is not the requester or the provider
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFoundResponse:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ServiceNotAvailableResponse:
      description: Service unavailable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UnexpectedErrorResponse:
      description: Unexpected error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  requestBodies:
    HelpRequestCreateRequest:
      description: Help request creation request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/HelpRequestCreate'
    HelpRequestUpdateRequest:
      description: Help request partial update request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/HelpRequestUpdate'
    HelpCreateRequest:
      description: Help creation request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/HelpCreate'
    HelpUpdateRequest:
      description: Help partial update request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/HelpUpdate'
  examples:
    BurningCatastropheHelpRequest:
      summary: A catastrophe at the stove - names a recipe, but neither a howToStep nor ingredients
      value: &burningCatastrophe
        helpRequestId: 23a8eeed-35f6-460b-892e-7bb458a8fded
        requester: f64e07f0-f9e7-4b7a-8695-e7a8b3ef5074
        title: Burning Catastrophe
        type: STEPS_TO_MITIGATE_CATASTROPHE
        description: Scones are burned and mother in law is coming in 30 minutes
        recipe: 7cf09822-77a1-46bb-812f-b7852bca0913
        preferredProvider:
          - GRANDMA_AVATAR
          - COMMUNITY
        status: OPEN
        createdAt: '2026-10-03T16:30:00Z'
        updatedAt: '2026-10-03T16:30:00Z'
    StayCalmHelp:
      summary: The Grandma Avatar's answer to the burning catastrophe - same type and recipe, no helpProvider
      value:
        helpId: a9caf90d-00b4-4a66-8184-7d02152e8d6a
        helpRequest: 23a8eeed-35f6-460b-892e-7bb458a8fded
        helpRequester: f64e07f0-f9e7-4b7a-8695-e7a8b3ef5074
        answerTitle: Stay calm
        helpProviderType: GRANDMA_AVATAR
        answer:
          answerType: STEPS_TO_MITIGATE_CATASTROPHE
          recipe: 7cf09822-77a1-46bb-812f-b7852bca0913
          explanation: use a new, cold pan
        createdAt: '2026-10-03T16:31:00Z'
        updatedAt: '2026-10-03T16:31:00Z'
    HelpRequestList:
      summary: One request per help type, each satisfying the invariants
      value:
        - *burningCatastrophe
        - helpRequestId: 0b6f3c1e-8f3a-4f0e-9a51-3c2d1e7f9a10
          requester: f64e07f0-f9e7-4b7a-8695-e7a8b3ef5074
          title: No buttermilk
          type: INGREDIENT_SUBSTITUTE
          description: What can I use instead of buttermilk?
          recipe: 7cf09822-77a1-46bb-812f-b7852bca0913
          ingredients:
            - 497f6eca-6276-4993-bfeb-53cbbbba6f08
          preferredProvider:
            - COMMUNITY
          status: OPEN
        - helpRequestId: 5e2d7a90-1c4b-4e8d-b3f6-9a0c2d4e6f81
          requester: f64e07f0-f9e7-4b7a-8695-e7a8b3ef5074
          title: Folding the dough
          type: PREPARATION_STEP_EXPLANATION
          description: What does "fold in" mean in step 3?
          recipe: 7cf09822-77a1-46bb-812f-b7852bca0913
          howToStep: 65610dee-fb83-4341-a730-26a4a99a621a
          preferredProvider:
            - GRANDMA_AVATAR
          status: ANSWERED
        - helpRequestId: 9c1a4e57-2b8d-4f63-a0e9-7d5b3c1f2e48
          requester: f64e07f0-f9e7-4b7a-8695-e7a8b3ef5074
          title: Dinner for my in-laws
          type: MENU_PROPOSAL
          description: Six guests on Saturday, one of them vegetarian
          preferredProvider:
            - CHEF
          status: OPEN
  securitySchemes:
    oAuth2:
      type: oauth2
      description: OAuth 2.0 Authorization Code Flow with PKCE
      flows:
        authorizationCode:
          authorizationUrl: https://larder.org/auth
          tokenUrl: https://larder.org/tokens
          refreshUrl: https://larder.org/refresh
          scopes:
            help:read: Grants read access to help requests and helps
            help:write: Grants write access to create, change and delete own help requests and helps